useSearchTypeFilter 훅 생성
